Move general classes from base.py to general.py
